--- src/main/java/org/side/med/config/WebClientConfig.java
@@ -1,14 +1,0 @@
-package org.side.med.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.reactive.function.client.WebClient;
-
-@Configuration
-public class WebClientConfig {
-
-    @Bean
-    public WebClient.Builder webClient() {
-        return WebClient.builder();
-    }
-}
