+++ src/main/java/org/side/med/controller/ApiController.java
@@ -2,7 +2,7 @@ package org.side.med.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.side.med.service.util.ApiServiceUtil;
+import org.side.med.util.ApiRestTemplateUtil;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -12,17 +12,17 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 @Slf4j
 public class ApiController {
-    private final ApiServiceUtil util;
+    private final ApiRestTemplateUtil util;
 
     @GetMapping("/search")
     public String searchDrug() {
-        log.info("ApiController.serachDrug");
+        log.info("ApiController.searchDrug");
         return "search Drug";
     }
 
     @GetMapping("/res")
     public String getKeyTest() {
         log.info("ApiController.resourceValueTest");
-        return util.serachDrug().toString();
+        return util.searchDrug();
     }
 }
